Create `utils.rs` in `tests-integration`
..to hold utility functions used internaly only.
